pyproject.toml
[project]
name = "photoprismproxy"
version = "0.1.0"
description = "Small web utility for more easily uploading photo(s) to PhotoPrism"
authors = [
    {name = "Radian LLC",email = "contact+photoprismproxy@radian.codes"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py310"]
